LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/access-reporting/property-opener-set.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: Blocked a frame with origin "http://web-platform.test:8800" from accessing a cross-origin frame. Protocols, domains, and ports must match.
CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: Blocked a frame with origin "http://web-platform.test:8800" from accessing a cross-origin frame. Protocols, domains, and ports must match.

PASS same-origin > w => w.opener = ""
PASS cross-origin > w => w.opener = ""
